+++ src/Facades/BybitWebSocket.php
@@ -1,0 +1,12 @@
+<?php
+namespace Tigusigalpa\ByBit\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+class BybitWebSocket extends Facade
+{
+    protected static function getFacadeAccessor()
+    {
+        return \Tigusigalpa\ByBit\BybitWebSocket::class;
+    }
+}